Fixed the loading screen in nexuiz a little
git-svn-id: https://svn.code.sf.net/p/fteqw/code/trunk@2558 fc73d0e0-1445-4013-8a0c-d673dee63da5

+++ engine/gl/gl_draw.c
@@ -945,8 +945,8 @@ TRACE(("dbg: GLDraw_ReInit: Allocating upload buffers\n"));
 	{
 		mpic_t *pic = Draw_SafeCachePic ("gfx/loading.lmp");
 		if (pic)
-			Draw_Pic ( (vid.width - pic->width)/2,
-				(vid.height - 48 - pic->height)/2, pic);
+			Draw_Pic ( ((int)vid.width - pic->width)/2,
+				((int)vid.height - 48 - pic->height)/2, pic);
 	}
 
 	TRACE(("dbg: GLDraw_ReInit: GL_EndRendering\n"));
